Added an executor for a step-wise performance test, and adjusted the result data structure for its needs (#23).

+++ jamopp.tests/src/test/java/tools/mdsd/jamopp/test/performance/stepwise/EvaluationStepFileChange.java
@@ -18,7 +18,7 @@ package tools.mdsd.jamopp.test.performance.stepwise;
 
 public class EvaluationStepFileChange {
 	private String path;
-	private int newSize;
+	private long newSize;
 
 	public String getPath() {
 		return path;
@@ -28,11 +28,11 @@ public class EvaluationStepFileChange {
 		this.path = path;
 	}
 	
-	public int getNewSize() {
+	public long getNewSize() {
 		return newSize;
 	}
 	
-	public void setNewSize(int newSize) {
+	public void setNewSize(long newSize) {
 		this.newSize = newSize;
 	}
 }
